Add functions / fill in functions for scoring curvatures
Strategy is to identify any "reasonably open" paths that have good clearance and allow the robot
to drive forward for a bit at max velocity. If there are reasonably open paths, we want to take the straightest one.

If there are not reasonably open paths, a variant of the scoring function from the lecture slides that is
based on free path length (not considering goal), clearance, and curvature for a particular curvature.

Testing:
* Untested except verified that ROS parameters can be used to change tunable parameters without recompiling